escc_journal_git_release/requirements.txt
numpy>=1.24,<3
pandas>=2.0,<3
scikit-learn>=1.3,<2
torch>=2.1
monai>=1.3
pyradiomics>=3.1
mamba-ssm>=2.2
